Revert "move evaluator, score_evaluator, test_metric back to the backend"
This reverts commit 344d2337ec96fa10f8e0bcc52515daccfcfd9b86. The evaluator should stay in the SDK
